# swagger.yaml
openapi: 3.0.3
info:
  title: Splitzies API
  description: |
    API for splitting receipts and assigning items to users.

    **Access model.** There are no accounts. `POST /devices` mints an anonymous
    device token, sent as `X-Device-Token`. Reading a bill needs only its ID,
    which is how a share link works. *Changing* a bill additionally requires
    that the calling device be on it — granted by uploading the receipt or by
    opening a live share link — so every mutating endpoint can answer `403
    not_a_participant`, and revoking a share link genuinely withdraws access.

    **Receipt photos are not stored.** The uploaded image is parsed in memory
    and discarded; only the extracted line items and receipt text are kept.

    **One device may keep 5 bills.** `POST /receipts/image` answers `409
    receipt_limit_reached` beyond that; deleting a bill frees a slot.
  version: 1.0.0
  contact:
    name: Splitzies
servers:
  - url: https://splitzies-backend-b74f565207ff.herokuapp.com
    description: Production
  - url: http://localhost:8080
    description: Local development
paths:
...
  /receipts/{receipt_id}/items:
    get:
      summary: Get items on a receipt
      description: List all line items on a receipt.
      operationId: getReceiptItems
      parameters:
        - $ref: '#/components/parameters/receipt_id'
      responses:
        '200':
          description: List of receipt items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetReceiptItemsResponse'
        '404':
          description: Receipt not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Add a unit to a receipt
      description: |
        Creates one individually-assignable unit. Pass group_id to add a unit
        to an existing group (the quantity stepper's "+"), or omit it to start
        a new group with this unit as its first member.
      operationId: createReceiptItem
      parameters:
        - $ref: '#/components/parameters/receipt_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReceiptItemRequest'
      responses:
        '201':
          description: Item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptItem'
        '400':
          description: Invalid request (missing name/amount when group_id is not set)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Receipt or group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or unknown X-Device-Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The calling device is not a participant on this bill
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /receipts/{receipt_id}/items/{item_id}:
    delete:
      summary: Delete a unit from a receipt
      description: |
        Deletes a single unit; its assignments are deleted with it. If it was
        the last unit in its group, the now-empty group is removed too.
      operationId: deleteReceiptItem
      parameters:
        - $ref: '#/components/parameters/receipt_id'
        - name: item_id
          in: path
          required: true
          schema:
            type: string
          description: Receipt item ID to delete
      responses:
        '200':
          description: Item deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Item not found on this receipt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or unknown X-Device-Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The calling device is not a participant on this bill
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /receipts/{receipt_id}/item-groups/{group_id}:
    patch:
      summary: Update a group's name and/or unit price
      description: |
        Updates apply to every unit in the group at once, since the review
        screen edits name and price per group rather than per unit.
      operationId: patchReceiptItemGroup
      parameters:
        - $ref: '#/components/parameters/receipt_id'
        - name: group_id
          in: path
          required: true
          schema:
            type: string
          description: Item group ID to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchReceiptItemGroupRequest'
      responses:
        '200':
          description: Item group updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid request (must include at least one of name or amount)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Item group not found on this receipt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or unknown X-Device-Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The calling device is not a participant on this bill
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    receipt_id:
      name: receipt_id
      in: path
      required: true
      schema:
        type: string
      description: Receipt ID (ULID)
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "receipt \"01J...\" not found"
        code:
          type: string
          description: Machine-readable error code
          example: "receipt_not_found"
        request_id:
          type: string
          description: Request ID from X-Request-Id header (when provided)
          example: "f5542cdd-b900-260c-591c-c771cf2df890"
    MoneyAmount:
      type: object
      description: Monetary amount with currency code and currency-aware decimal precision.
      properties:
        amount:
          type: number
          format: double
          description: Numeric value rounded to the currency's decimal places (e.g. 2 for USD, 0 for JPY).
          example: 12.95
        currency:
          type: string
          description: ISO 4217 currency code.
          example: USD
      example:
        amount: 12.95
        currency: USD
    ReceiptItem:
      type: object
      description: |
        A single, individually-assignable unit. Items sharing group_id are
        grouped for display only (e.g. 3 "Margarita" units) — assignment and
        splitting always happen per item, never per group.
      properties:
        id:
          type: string
          description: Item ID (ULID)
        group_id:
          type: string
          description: Groups units that display together (e.g. quantity 3 of the same item)
        group_name:
          type: string
          description: Display name of the group
        name:
          type: string
          description: Item name
        display_order:
          type: integer
          description: Sort order within the receipt
        amount:
          $ref: '#/components/schemas/MoneyAmount'
          nullable: true
          description: Per-unit price
...
    GetReceiptItemsResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReceiptItem'
...
    CreateReceiptItemRequest:
      type: object
      description: |
        When group_id is set, the new unit joins that existing group and
        adopts its name and price — name/amount are ignored. When group_id is
        omitted, name and amount are required and a new group is created.
      properties:
        name:
          type: string
          description: Required when group_id is not set
          example: "Margarita"
        amount:
          type: number
          format: double
          description: Required when group_id is not set
          example: 12.00
        group_id:
          type: string
          nullable: true
          description: Existing group to add this unit to
    PatchReceiptItemGroupRequest:
      type: object
      description: |
        At least one of name or amount must be provided. Applies to every unit
        in the group — the review screen edits a group's name and price as one,
        not per unit.
      minProperties: 1
      properties:
        name:
          type: string
          nullable: true
        amount:
          type: number
          format: double
          nullable: true
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: "share link revoked"
...